+++ src/app/app-routing.module.ts
@@ -1,11 +1,9 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { HomeComponent } from './pages/home/home.component';
-import { URLS } from './const/urls';
 
-const routes: Routes = [
-  { path: URLS.HOME, component: HomeComponent, data: { depth: 0 } },
-];
+
+const routes: Routes = [];
+
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
   exports: [RouterModule]
