+++ src/app/app.component.scss
@@ -15,7 +15,7 @@ $breakpoint-lg: 1360px;
   }
 }
 
-.route-container {
+.context {
   widows: 100vw;
   display: flex;
   justify-content: center;
