+++ src/app/app.component.ts
@@ -1,5 +1,4 @@
 import { Component } from '@angular/core';
-import { RouterOutlet } from '@angular/router';
 
 @Component({
   selector: 'app-root',
@@ -8,8 +7,4 @@ import { RouterOutlet } from '@angular/router';
 })
 export class AppComponent {
   title = 'proyectoPoli';
-
-  getDepth(outlet: RouterOutlet) {
-    return outlet.activatedRouteData.depth;
-  }
 }
