+++ src/app/app.module.ts
@@ -6,18 +6,16 @@ import { AppComponent } from './app.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { GraphQLModule } from './graphql.module';
 import { HttpClientModule } from '@angular/common/http';
-import { HomeComponent } from './pages/home/home.component';
+import { HomeComponent } from './home/home.component';
 import { MaterialModule } from './marerial-module';
-import { PrincipalMenuComponent } from './pages/principal-menu/principal-menu.component';
-import { HeaderComponent } from './components/header/header.component';
+import { PrincipalMenuComponent } from './principal-menu/principal-menu.component';
 
 
 @NgModule({
   declarations: [
     AppComponent,
     HomeComponent,
-    PrincipalMenuComponent,
-    HeaderComponent,
+    PrincipalMenuComponent
   ],
   imports: [
     BrowserModule,
