--- src/app/components/header/header.component.html
@@ -1,3 +1,0 @@
-<mat-card>
-  <h1>Proyecto Politecnico</h1>
-</mat-card>
