--- src/app/const/members.ts
@@ -1,24 +1,0 @@
-import { IMembers } from '../models/member-model';
-
-export const Members: IMembers[] = [
-  {
-    id: '1911981574',
-    name: 'Faber Duvan Herrera Martinez'
-  },
-  {
-    id: '1911980714',
-    name: 'Wendy Vanessa Pedroza Aguirre'
-  },
-  {
-    id: '1911981307',
-    name: 'Leonardo Castiblanco ochoa'
-  },
-  {
-    id: '1710650420',
-    name: 'Paulo Leal'
-  },
-  {
-    id: '0921020173',
-    name: 'Diego Montealegre Rozo'
-  },
-];
