--- src/app/const/urls.ts
@@ -1,3 +1,0 @@
-export const URLS = {
-  HOME: '',
-};
