+++ src/app/home/home.component.html
@@ -1,4 +1,6 @@
-<app-header></app-header>
+<mat-card>
+  <h1>Proyecto Politecnico</h1>
+</mat-card>
 <mat-card>
   <mat-list>
     <mat-list-item>
