+++ src/app/models/member-model.ts
@@ -2,3 +2,26 @@ export interface IMembers {
   name: string;
   id: string;
 }
+
+export const Members: IMembers[] = [
+  {
+    id: '1911981574',
+    name: 'Faber Duvan Herrera Martinez'
+  },
+  {
+    id: '1911980714',
+    name: 'Wendy Vanessa Pedroza Aguirre'
+  },
+  {
+    id: '1911981307',
+    name: 'Leonardo Castiblanco ochoa'
+  },
+  {
+    id: '1710650420',
+    name: 'Paulo Leal'
+  },
+  {
+    id: '0921020173',
+    name: 'Diego Montealegre Rozo'
+  },
+];
