+++ src/app/principal-menu/principal-menu.component.scss
