+++ src/app/principal-menu/principal-menu.component.spec.ts
